Move stylus support from compile time to run time.
Builds which use stylus or touch input can #define STYLUS_BASED, to make
input more convenient if there is only one button available. As this is
a compile time option, the same build could not be adapted for mouse and
touch input simultaneously.

Add a new game flag bit, STYLUS_SUPPORT, which asserts that a game
handles stylus input specially, and a new modifier, MOD_STYLUS, which
asserts that this mouse event originates from a stylus or touch.

This will allow games to handle mouse and stylus input differently in
the same build.