GTK: emulate keys with buttons for touch devices.
Detect if a touchscreen device is active, and show a row of buttons
which emulate the required keys.

For games which use both mouse buttons, a toggle button swaps short and
long press. To avoid confusion, if a mouse is also present, left and
right click are swapped too.

Undo and Redo buttons are also shown on the keybar.

The keybar must be able to wrap onto multiple lines, to support games
which require many keys on devices with small screens. It looks tidier
if the buttons all have the same size. This is managed by the WrapBox
component.

The common buttons are assigned a standard GTK icon. For consistency and
to save space, all keybar buttons are styled as icon buttons, even if
they contain text.