Mosaic: don't emit invalid '}' in game descriptions.
If a game description involved a run of 27 clue-free cells, then the
game would accidentally emit '}', obtained by incrementing the 'z' for
a 26-space run one more time, due to an off-by-one error. Not only
that, but the grid would be incorrectly encoded even if you did
interpret '}' as a 27-space run, because that character was emitted
_instead_ of incrementing the count, not as well.

Now a run of 27 spaces is emitted as 'za', and 28 as 'zb', as it
should be.

A particular game seed which ran into this was 50x50h0#524000234917895
which produced a game description string that validate_desc wouldn't
accept. Worse, if you entered that seed live, so that the string
didn't go through validate_desc at all, then it would be mis-decoded,
moving some 7 clues on to the left margin where they're obviously
unsatisfiable.